requirements.txt
# Python dependencies for IBW Virtual Advisor
# Python version: Python 3.13.0
# Generated on: Fri Jan 31 02:11:04 CET 2025

annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.2.1
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
fastapi==0.115.8
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2024.12.0
google-auth==2.38.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
jiter==0.8.2
kubernetes==32.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
numpy==2.2.2
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.60.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
posthog==3.11.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.21.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
zipp==3.21.0
